refactor: extract package materializer
